+++ src/main/java/kitchenpos/bo/TableGroupBo.java
@@ -1,9 +1,9 @@
 package kitchenpos.bo;
 
-import kitchenpos.dao.OrderDao;
+import kitchenpos.orders.dao.OrderDao;
 import kitchenpos.dao.OrderTableDao;
 import kitchenpos.dao.TableGroupDao;
-import kitchenpos.model.OrderStatus;
+import kitchenpos.orders.model.OrderStatus;
 import kitchenpos.model.OrderTable;
 import kitchenpos.model.TableGroup;
 import org.springframework.stereotype.Component;
