+++ src/main/java/kitchenpos/orders/model/Order.java
@@ -1,4 +1,6 @@
-package kitchenpos.model;
+package kitchenpos.orders.model;
+
+import kitchenpos.model.OrderLineItem;
 
 import java.time.LocalDateTime;
 import java.util.List;
