+++ src/main/java/kitchenpos/orders/model/OrderStatus.java
@@ -1,4 +1,4 @@
-package kitchenpos.model;
+package kitchenpos.orders.model;
 
 public enum OrderStatus {
     COOKING, MEAL, COMPLETION
